Move line storage builder
